Create Enemy Goon HealthBar

+++ Breadth of the Wild Goose Project/Assets/Scripts/Extra Scripts/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void ApplyDamage(float amount);
+    void ApplyDamage(int amount);
 }
